// product/src/modules/grpc/proto/category.pb.ts
export interface Banner {
  lang: string;
  url: string;
  source: string;
}

export interface PriceQuality {
  name: string;
}

export interface Condition {
  id: string;
  name: string;
  nameAr: string;
  labelColor: string;
  textColor: string;
  banners: Banner[];
  positionEn: number;
  positionAr: number;
}